change the view type of pixel raw data as float

+++ carta/cpp/testProtoBuf/CartaHelper.h
@@ -10,13 +10,13 @@ Carta::Lib::NdArray::RawViewInterface* getRawData(std::shared_ptr<Carta::Lib::Im
 std::vector<double> getHertzValues(std::shared_ptr<Carta::Lib::Image::ImageInterface> m_image,
                                    std::vector<int> dims, int spectralIndex);
 
-std::vector<double> extractRawData(std::shared_ptr<Carta::Lib::Image::ImageInterface> astroImage,
+std::vector<float> extractRawData(std::shared_ptr<Carta::Lib::Image::ImageInterface> astroImage,
                                    int frameStart, int frameEnd, int stokeIndex, int mip);
 
-bool downVector(std::vector<double> &rawData, int x_size, int y_size, int mip);
+bool downVector(std::vector<float> &rawData, int x_size, int y_size, int mip);
 
-std::vector<double> downSampling(Carta::Lib::NdArray::RawViewInterface *view, int ilb, int iub, int jlb, int jub, int mip);
+std::vector<float> downSampling(Carta::Lib::NdArray::RawViewInterface *view, int ilb, int iub, int jlb, int jub, int mip);
 
-void compareVectors(std::vector<double> vec1, std::vector<double> vec2);
+void compareVectors(std::vector<float> vec1, std::vector<float> vec2);
 
 #endif
