+++ carta/cpp/testProtoBuf/ProtoBufHelper.h
@@ -1,7 +1,7 @@
 #ifndef ProtoBuffHelper_h
 #define ProtoBuffHelper_h
 
-void PromptForRasterImageData(Stream::RasterImageData* RasterImageData, const std::vector<double> data);
+void PromptForRasterImageData(Stream::RasterImageData* RasterImageData, const std::vector<float> data);
 
 void ListData(const Stream::DataBook& DataBook);
